+++ Phoenix/include/Warping.h
@@ -14,14 +14,14 @@ namespace Warping
 {
     std::vector<earcutPoly> getPolygonsInBox(const Library& lib, const Polygon* box);
 
-    Eigen::Matrix3d getPerspectiveMatrixTransform(const double2 src[4], const double2 dst[4]);
+    Eigen::Matrix3d getPerspectiveMatrixTransform(const float2 src[4], const float2 dst[4]);
 
     void applyMatrixToPolygons(Eigen::Matrix3d& m, std::vector<earcutPoly>& polys);
 
     // boite représentant la boite cible et sa destination réelle
     struct Boxes {
-        double2 src[4];
-        double2 dst[4];
+        float2 src[4];
+        float2 dst[4];
     };
 }
 
@@ -30,14 +30,14 @@ static std::vector<Warping::Boxes> src_dst_boxes =
 {
     // 1ere transformation
     Warping::Boxes{
-        double2{0.0, 0.0},
-        double2{0.0, 20000.0},
-        double2{20000.0, 20000.0},
-        double2{20000.0, 0.0},
+        float2{0, 0},
+        float2{0, 20000},
+        float2{20000, 20000},
+        float2{20000, 0},
 
-        double2{1000.0, 500.0},
-        double2{300.0, 10000.0},
-        double2{15000.0, 15000.0},
-        double2{18000.0, 1000.0}
+        float2{1000, 500},
+        float2{300, 10000},
+        float2{15000, 15000},
+        float2{18000, 1000}
     }
 };
